get rid of mips-specific types.h since common one contains more fixes

+++ libc/sysdeps/linux/mips/bits/kernel_types.h
@@ -14,7 +14,7 @@ typedef unsigned int	__kernel_mode_t;
 typedef unsigned int	__kernel_nlink_t;
 typedef long		__kernel_off_t;
 typedef int		__kernel_pid_t;
-typedef int		__kernel_ipc_pid_t;
+typedef long int		__kernel_ipc_pid_t;
 typedef int		__kernel_uid_t;
 typedef int		__kernel_gid_t;
 typedef unsigned long	__kernel_size_t;
@@ -50,7 +50,7 @@ typedef unsigned long  __kernel_nlink_t;
 
 typedef long		__kernel_off_t;
 typedef int		__kernel_pid_t;
-typedef int		__kernel_ipc_pid_t;
+typedef long int		__kernel_ipc_pid_t;
 typedef int		__kernel_uid_t;
 typedef int		__kernel_gid_t;
 typedef unsigned int	__kernel_size_t;
